+++ ExpandWorldPrefabs/Commands.cs
@@ -1,55 +1,52 @@
+using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
+using Data;
 using Service;
-using UnityEngine;
 
 namespace ExpandWorld.Prefab;
 
 public class Commands
 {
 
-  public static void Run(Info info, ZDO? zdo, Dictionary<string, string> parameters, List<PlayerInfo> players)
+  public static void Run(Info info, ZDO? zdo, Dictionary<string, string> parameters)
   {
     if (info.Commands.Length == 0) return;
     var pars = parameters;
-    // PlayerSearch replaces the <pchar>, <pid>, and <pname> parameters.
-    if (info.PlayerSearch != PlayerSearch.None)
-      pars = parameters.Where(p => p.Key != "<pchar>" && p.Key != "<pid>" && p.Value != "<pname>").ToDictionary(p => p.Key, p => p.Value);
     var commands = info.Commands.Select(s => Helper.ReplaceParameters(s, parameters, zdo)).ToArray();
-    if (info.PlayerSearch == PlayerSearch.None)
-      CommandManager.Run(commands, players.Count == 0 ? null : players[0]);
-    else
-      CommandManager.Run(commands, players);
+    Run(commands);
   }
-  public static void Run(Info info, Dictionary<string, string> parameters, List<PlayerInfo> players) =>
-    Run(info, null, parameters, players);
-  public static List<PlayerInfo> FindPlayers(ZDO zdo, ZDO? source, Info info)
-  {
-    var players = FindPlayers();
-    if (info.PlayerSearch == PlayerSearch.None)
-      return players.Where(p => p.ZDOID == zdo.m_uid || p.ZDOID == source?.m_uid).ToList();
+  public static void Run(Info info, Dictionary<string, string> parameters) =>
+    Run(info, null, parameters);
 
-    var pos = zdo.m_position;
-    var rot = zdo.m_rotation;
-    players = players.Where(p =>
-      Utils.DistanceXZ(p.Pos, pos) <= info.PlayerSearchDistance
-      && (info.PlayerSearchHeight == 0f || Mathf.Abs(p.Pos.y - pos.y) <= info.PlayerSearchHeight
-    )).ToList();
-    if (info.PlayerSearch == PlayerSearch.All)
-      return players;
-    if (info.PlayerSearch == PlayerSearch.Closest)
+  private static void Run(IEnumerable<string> commands)
+  {
+    var parsed = commands.Select(c => Parse(c)).ToArray();
+    foreach (var cmd in parsed)
     {
-      var player = players.OrderBy(p => Utils.DistanceXZ(p.Pos, pos)).FirstOrDefault();
-      return player == null ? [] : [player];
+      try
+      {
+        Console.instance.TryRunCommand(cmd);
+      }
+      catch (Exception e)
+      {
+        Log.Error($"Failed to run command: {cmd}\n{e.Message}");
+      }
     }
-    return [];
   }
-  public static List<PlayerInfo> FindPlayers()
+  private static string Parse(string command)
   {
-    var players = ZNet.instance.GetPeers().Select(p => new PlayerInfo(p)).ToList();
-    if (ZNet.instance.IsServer() && !ZNet.instance.IsDedicated())
-      players.Add(new(Player.m_localPlayer));
-
-    return players;
+    var expressions = command.Split(' ').Select(s => s.Split('=')).Select(a => a[a.Length - 1].Trim()).SelectMany(s => s.Split(',')).ToArray();
+    foreach (var expression in expressions)
+    {
+      if (expression.Length == 0) continue;
+      // Single negative number would get handled as expression.
+      var sub = expression.Substring(1);
+      if (!sub.Contains('*') && !sub.Contains('/') && !sub.Contains('+') && !sub.Contains('-')) continue;
+      var value = Calculator.EvaluateFloat(expression) ?? 0f;
+      command = command.Replace(expression, value.ToString("0.#####", NumberFormatInfo.InvariantInfo));
+    }
+    return command;
   }
 }