+++ ExpandWorldPrefabs/RpcInfo.cs
@@ -58,7 +58,7 @@ public abstract class RpcInfo
       Delay = Parse.Float(d, 0f);
     Parameters = lines.OrderBy(p => p.Key).Where(p => Parse.TryInt(p.Key, out var _)).Select(p => Parse.Kvp(p.Value)).ToArray();
   }
-  public void Invoke(ZDO zdo, Dictionary<string, string> parameters, List<PlayerInfo>? players)
+  public void Invoke(ZDO zdo, Dictionary<string, string> parameters)
   {
     var source = ZRoutedRpc.instance.m_id;
     if (SourceParameter != null)
@@ -78,24 +78,12 @@ public abstract class RpcInfo
       if (peerId.HasValue)
         DelayedRpc.Add(Delay, source, peerId.Value, GetId(zdo), Hash, pars);
     }
-    else if (Target == RpcTarget.Search && players != null)
-    {
-      // Probably could inject some player specific parameters here.
-      foreach (var player in players)
-        DelayedRpc.Add(Delay, source, player.PeerId, GetId(zdo), Hash, pars);
-    }
   }
-  public void InvokeGlobal(Dictionary<string, string> parameters, List<PlayerInfo>? players)
+  public void InvokeGlobal(Dictionary<string, string> parameters)
   {
     var source = ZRoutedRpc.instance.m_id;
     var pars = Packaged ? PackagedGetParameters(parameters) : GetParameters(parameters);
-    if (Target == RpcTarget.Search && players != null)
-    {
-      // Probably could inject some player specific parameters here.
-      foreach (var player in players)
-        DelayedRpc.Add(Delay, source, player.PeerId, ZDOID.None, Hash, pars);
-    }
-    else DelayedRpc.Add(Delay, source, ZRoutedRpc.Everybody, ZDOID.None, Hash, pars);
+    DelayedRpc.Add(Delay, source, ZRoutedRpc.Everybody, ZDOID.None, Hash, pars);
   }
   private object[] GetParameters(ZDO? zdo, Dictionary<string, string> parameters)
   {